requirements.txt
python-dotenv
playsound
requests
langdetect
transformers
torch
flask